back fix unauthorized delete post by user

+++ src/api/build/controllers/post.controller.js
@@ -97,10 +97,22 @@ let PostController = class PostController extends paradigm_express_webapi_1.ApiC
     deletePost(postId) {
         return __awaiter(this, void 0, void 0, function* () {
             try {
-                yield this.repoComment.deleteComments(parseInt(this.httpContext.request.params.postId));
-                yield this.repoValoration.deleteValorations(parseInt(this.httpContext.request.params.postId));
-                const data = yield this.repoPost.delete(parseInt(this.httpContext.request.params.postId));
-                this.httpContext.response.status(200).send(data);
+                const { gameId, postId } = this.httpContext.request.params;
+                const getPostById = yield this.repoPost.find("game_id = ? AND id = ?", [
+                    gameId,
+                    postId,
+                ]);
+                console.log("getPostById[0]", getPostById[0].created_by); // obtener de acá el id del usuario que creó el post
+                const { created_by } = this.httpContext.request.body; // este que el id que mando desde el token del front 
+                if (created_by == getPostById[0].created_by) {
+                    yield this.repoComment.deleteComments(parseInt(this.httpContext.request.params.postId));
+                    yield this.repoValoration.deleteValorations(parseInt(this.httpContext.request.params.postId));
+                    const data = yield this.repoPost.delete(parseInt(this.httpContext.request.params.postId));
+                    this.httpContext.response.status(200).send(data);
+                }
+                else {
+                    this.httpContext.response.status(403).send("Not authorized to do this action");
+                }
             }
             catch (error) {
                 console.log(error);
